Introduce EventStreamProcessing module
