+++ lib/eventory.rb
@@ -3,16 +3,10 @@ require 'eventory/version'
 require 'eventory/event_data'
 require 'eventory/event'
 require 'eventory/recorded_event'
-require 'eventory/subscription'
-require 'eventory/persistent_subscription'
-require 'eventory/checkpoints'
-require 'eventory/checkpoint'
 require 'eventory/event_handlers'
-require 'eventory/event_stream_processor'
-require 'eventory/projector'
-require 'eventory/reactor'
 require 'eventory/domain'
 require 'eventory/event_store'
+require 'eventory/event_stream_processing'
 
 module Eventory
 end
