+++ lib/eventory/event_stream_processing.rb
@@ -1,0 +1,12 @@
+require 'eventory/event_stream_processing/postgres/checkpoint'
+require 'eventory/event_stream_processing/postgres/checkpoints'
+require 'eventory/event_stream_processing/event_stream_processor'
+require 'eventory/event_stream_processing/persistent_subscription'
+require 'eventory/event_stream_processing/projector'
+require 'eventory/event_stream_processing/reactor'
+require 'eventory/event_stream_processing/subscription'
+
+module Eventory
+  module EventStreamProcessing
+  end
+end
