+++ lib/eventory/event_stream_processing/projector.rb
@@ -1,0 +1,6 @@
+module Eventory
+  module EventStreamProcessing
+    class Projector < EventStreamProcessor
+    end
+  end
+end
