+++ lib/eventory/event_stream_processing/reactor.rb
@@ -1,0 +1,6 @@
+module Eventory
+  module EventStreamProcessing
+    class Reactor < EventStreamProcessor
+    end
+  end
+end
