--- lib/eventory/projector.rb
@@ -1,4 +1,0 @@
-module Eventory
-  class Projector < EventStreamProcessor
-  end
-end
