--- lib/eventory/reactor.rb
@@ -1,4 +1,0 @@
-module Eventory
-  class Reactor < EventStreamProcessor
-  end
-end
